Mejora de rendimiento

+++ include/Math/types/Vec2.hpp
@@ -11,7 +11,7 @@ namespace EasySDL
         Vec2() : x(0), y(0) {}
         Vec2(float x, float y) : x(x), y(y) {}
 
-        int distance(const Vec2 &other)
+        float distance(const Vec2 &other)
         {
             return sqrt(pow((other.x - this->x), 2) + pow((other.y - this->y), 2));
         }
